% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_roxygen.R
\name{as_roxygen_docs}
\alias{as_roxygen_docs}
\title{Generate Standalone \code{?options} Documentation}
\usage{
as_roxygen_docs(
  title = paste(pkgname(env), "Options"),
  desc = default_options_rd_desc(),
  env = parent.frame()
)
}
\arguments{
\item{title}{An optional, customized title (defaults to "Options")}

\item{desc}{An optional, customized description of behaviors}

\item{env}{An environemnt in which to find the associated options object}
}
\value{
A character vector of \code{roxygen2} tag segments
}
\description{
Produce a comprehensive documentation page outlining all your defined
options' behaviors.
}
\examples{
#' @eval options::as_roxygen_docs()
NULL

}
\seealso{
Other options_roxygen2: 
\code{\link{as_params}()}
}
\concept{options_roxygen2}
\keyword{roxygen2}
