% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{assert_naming_fn_signature}
\alias{assert_naming_fn_signature}
\title{Assert signature for naming functions}
\usage{
assert_naming_fn_signature(fn)
}
\arguments{
\item{fn}{A function to inspect}
}
\description{
Assert signature for naming functions
}
\keyword{internal}
