% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R, R/options_get.R
\name{envvar_fns}
\alias{envvar_fns}
\alias{envvar_parse}
\alias{envvar_is_true}
\alias{envvar_is_set}
\alias{envvar_is_false}
\alias{envvar_str_split}
\title{Utility functions for processing option values from strings}
\usage{
envvar_parse(...)

envvar_is_true(...)

envvar_is_set(...)

envvar_is_false(...)

envvar_str_split(delim = ";", ...)

envvar_parse(...)

envvar_is_true(...)

envvar_is_set(...)

envvar_is_false(...)

envvar_str_split(delim = ";", ...)
}
\description{
Utility functions for processing option values from strings

Utility functions for processing option values from strings
}
