% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err}
\alias{err}
\title{Raise a package error}
\usage{
err(title, issues = list(), which = 0)
}
\arguments{
\item{title}{A title for the error}

\item{issues}{An optionally named list of issues to associate with the error.
When named, the issues are first sorted by issue name.}

\item{which}{A relative frame to use to build the associated call}
}
\value{
An options_error class
}
\description{
Raise a package error
}
\keyword{internal}
