% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_spec.R
\name{format.option_spec}
\alias{format.option_spec}
\title{Format an option specification}
\usage{
\method{format}{option_spec}(x, value, ..., fmt = options_fmts())
}
\arguments{
\item{x}{An option specification ("option_spec") class object}

\item{value}{Optionally, the current value to display for the option being
specified}

\item{...}{Additional arguments unused}

\item{fmt}{A list of formats to use for formatting individual text elements}
}
\value{
A formatted character value
}
\description{
Format an option specification
}
\keyword{internal}
