% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_env.R
\name{format.options_env}
\alias{format.options_env}
\title{Format an options environment}
\usage{
\method{format}{options_env}(x, ..., fmt = options_fmts())
}
\arguments{
\item{x}{An option environment ("option_env") class object}

\item{...}{Additional arguments unused}

\item{fmt}{A list of formats to use for formatting individual text elements}
}
\value{
A formatted character value
}
\description{
Format an options environment
}
\keyword{internal}
