% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_env.R
\name{format.options_list}
\alias{format.options_list}
\title{Format an options list}
\usage{
\method{format}{options_list}(x, ..., fmt = options_fmts())
}
\arguments{
\item{x}{An option list ("option_list") class object}

\item{...}{Additional arguments unused}

\item{fmt}{A list of formats to use for formatting individual text elements}
}
\value{
A formatted character value
}
\description{
Format an options list
}
\keyword{internal}
