% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_spec.R
\name{format_field}
\alias{format_field}
\title{Format a possible option source}
\usage{
format_field(field, active, value, fmt = options_fmts())
}
\arguments{
\item{field}{The field for the option source}

\item{active}{Whether this source is the source of the option's value}

\item{value}{The value from this source that was used}

\item{fmt}{A list of formats to use for formatting individual text elements}
}
\value{
A formatted character value
}
\description{
Format a possible option source
}
\keyword{internal}
