% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_spec.R
\name{format_value}
\alias{format_value}
\alias{format_value.default}
\alias{format_value.S3}
\alias{format_value.S4}
\alias{format_value.function}
\alias{format_value.environment}
\alias{format_value.expression}
\alias{format_value.quote}
\alias{format_value.call}
\alias{format_value.name}
\alias{format_value.symbol}
\title{Format value shorthands for command line display}
\usage{
format_value(x, ..., fmt = NULL)

\method{format_value}{default}(x, ..., fmt = options_fmts())

\method{format_value}{S3}(x, ..., fmt = options_fmts())

\method{format_value}{S4}(x, ..., fmt = options_fmts())

\method{format_value}{`function`}(x, ..., fmt = options_fmts())

\method{format_value}{environment}(x, ..., fmt = options_fmts())

\method{format_value}{expression}(x, ..., fmt = options_fmts())

\method{format_value}{quote}(x, ..., fmt = options_fmts())

\method{format_value}{call}(x, ..., fmt = options_fmts())

\method{format_value}{name}(x, ..., fmt = options_fmts())

\method{format_value}{symbol}(x, ..., fmt = options_fmts())
}
\arguments{
\item{x}{An R object to display, attempting to show the actual value, but
falling back to shorthands for more complex data types.}

\item{...}{Additional arguments unused}

\item{fmt}{A list of formats to use for formatting individual text elements}
}
\value{
A formatted character value
}
\description{
Format value shorthands for command line display
}
\keyword{internal}
