% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_env.R
\name{get_options_env}
\alias{get_options_env}
\title{Fetch options environment from package namespace}
\usage{
get_options_env(env)
}
\arguments{
\item{env}{An environment in which to search for an options environment}
}
\value{
an options environment
}
\description{
Initialize (if needed) and retrieve an environment containing options
}
\keyword{internal}
