% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getOption.R
\name{getOption}
\alias{getOption}
\title{getOption - dropin replacement}
\usage{
getOption(x, default = NULL)
}
\arguments{
\item{x}{unquoted expression for retrieving the option}

\item{default}{if the specified option is not set in the options list, this
value is returned. This facilitates retrieving an option and checking whether
 it is set and setting it separately if not.}
}
\value{
The current value for option \code{x} or \code{NULL} if unset.
}
\description{
\code{getOption} allows for deep-inspection of Global Options. The problem
with the base implementation of \code{getOption} is that retrieves options
by name. This make is difficult to work with deeply layered options and
promotes package developers to create top-level global options like
\code{data.table.xxx} rather than using \code{data.table$xxx}
}
\details{
Dropin replacement for \code{getOption} that allows for deep retrieval rather
than access by name alone
}
\examples{
options( foo = list( bar="baz" ) )

   getOption( 'foo' )
   getOption( foo )
   getOption( foo$bar )       # "baz"
   getOption( foo[['bar']] )  # "baz"

   options( lets = letters )
   getOption( lets[1:5] )
}
\seealso{
\code{\link[base]{getOption}} \cr
  \code{\link{setOption}}
}

