% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{naming}
\alias{naming}
\alias{set_envvar_name_fn}
\alias{set_option_name_fn}
\title{Define Naming Conventions}
\usage{
set_envvar_name_fn(fn, env = parent.frame())

set_option_name_fn(fn, env = parent.frame())
}
\arguments{
\item{fn}{A callback function which expects two arguments, the package name
and option name, and returns a single character value to use as an
environment variable name.}

\item{env}{An environment in which to search for options settings}
}
\value{
fn
}
\description{
Define Naming Conventions
}
\section{Functions}{
\itemize{
\item \code{set_envvar_name_fn()}: Set a callback function to use to format environment variable names.

\item \code{set_option_name_fn()}: Set a callback function to use to format option names.

}}
