% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_get.R
\name{opt}
\alias{opt}
\title{Retrieve option}
\usage{
opt(x, default = NULL, env = parent.frame())
}
\arguments{
\item{x}{an option name}

\item{default}{a default value if the option is not set}
}
\description{
Retrieve option
}
