% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayon.R
\name{options_fmts}
\alias{options_fmts}
\title{Optional Crayon Handling}
\usage{
options_fmts()
}
\value{
formatted text
}
\description{
Generate a list of styling functions using \code{crayon}, while safely falling
back to non-\code{crayon} output when \code{crayon} is unavailable.
}
\keyword{internal}
