% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_get.R
\name{opts}
\alias{opts}
\title{Fetch options}
\usage{
opts(xs, env = parent.frame())
}
\arguments{
\item{env}{An environment in which to search for options}
}
\value{
the options environment as a list
}
\description{
Retrieve options as list from namespace
}
