% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgname}
\alias{pkgname}
\title{Grab package name, at runtime}
\usage{
pkgname(env = parent.frame())
}
\description{
Lazily grab \code{packageName()} within calling environment, not within function
environment.
}
