% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_option.R
\name{reflow_option_desc}
\alias{reflow_option_desc}
\title{Reflow multiline strings}
\usage{
reflow_option_desc(x)
}
\arguments{
\item{x}{A vector of multiline strings to reflow}
}
\value{
The reflowed strings
}
\description{
A small helper function for allowing multiline strings to be collapsed into
continuous lines, similar to markdown's paragraph handling.
}
\keyword{internal}
