% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/set_option.R
\name{set_option}
\alias{set_option}
\title{set_option}
\usage{
set_option(...)
}
\arguments{
\item{...}{named arguments passed to \code{options} -or- a single named list.

Base has a \code{getOption} function, but does not have the complimentary
\code{setOption} function.  The \code{options} function is used instead.
\code{set_option} aliases for \code{options}}
}
\value{
If an argument is provided, behavior is as \code{\link[base]{options}}:
  For \code{options(name)}, a list of length one containing the set value,
  or NULL if it is unset. For uses setting one or more options, a list with
  the previous values of the options changed (returned invisibly).

  If no argument is provided the \code{NULL} is returned invisibly. This is
  different from \code{\link[base]{options}}.
}
\description{
An alias for \code{options}
}
\examples{
set_option( foo = list(bar='baz') )

  set_option()  # NULL, Non-op

  # NOT IMPLEMENTED YET:
  \dontrun{
    set_option( foo$bar := 'baz' )
    set_option( foo$bar ~ 'baz' )
  }
}
\seealso{
\code{\link[base]{options}}
}

