% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/setOption.R
\name{setOption}
\alias{setOption}
\title{setOption}
\usage{
setOption(...)
}
\arguments{
\item{...}{named arguments passed to \code{options} -or- a single named list.

It is surprising that base has a \code{getOption} function, but does not have
the complimentary \code{setOption} function and uses \code{options} instead.
\code{setOption} is an alias for \code{options}}
}
\description{
An alias for \code{options}
}
\examples{
setOption( foo = list(bar='baz') )
}
\seealso{
\code{\link[base]{options}}
}

