% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_env.R
\name{get_options_env}
\alias{get_options_env}
\alias{get_options_env.options_env}
\alias{get_options_env.options_list}
\alias{get_options_env.default}
\title{Retrieve options environment (experimental)}
\usage{
get_options_env(env, ...)

\method{get_options_env}{options_env}(env, ...)

\method{get_options_env}{options_list}(env, ...)

\method{get_options_env}{default}(
  env = parent.frame(),
  ...,
  inherits = FALSE,
  ifnotfound = emptyenv()
)
}
\arguments{
\item{env}{An environment in which to search for an options environment}

\item{...}{Additional arguments unused}

\item{inherits}{Whether to search upward through parent environments}

\item{ifnotfound}{A result to return of no options environment is found.}
}
\value{
An environment containing option specifications and default values,
or \code{ifnotfound} if no environment is found.
}
\description{
The options environment stores metadata regarding the various options
defined in the local scope - often the top environment of a package
namespace.
}
\note{
This function's public interface is still under consideration. It is
surfaced to provide access to option names, though the exact mechanism
of retrieving these names should be considered experimental.
}
