% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgname}
\alias{pkgname}
\title{Grab package name, at runtime}
\usage{
pkgname(env = parent.frame())
}
\arguments{
\item{env}{An environment in which to search for a package name}
}
\value{
A package name or "globalenv" if not found
}
\description{
Lazily grab \code{packageName()} within calling environment, not within function
environment.
}
\keyword{internal}
